doc: Move sections about structured reply values
Most of the documentation follows the pattern of a high-level
overview in the first half, then details about specific values
to use in the second half.  To fit in, structured replies should
place the details about flags and chunk types alongside other
transmission phase constants.

While at it, tweak a couple of wording issues outside of the
section shuffling.

Signed-off-by: Eric Blake <eblake@redhat.com>
Signed-off-by: Alex Bligh <alex@alex.org.uk>