Fix compilation errors on old glib versions
On centos6 the older version of glib needs to have g_thread_init() run before we
create the threadpool, however this is a deprecated function and we don't need
it on newer glibs.  So add a configure check to see if our glib is the old one
and add the appropriate includes and g_thread_init() call.  Tested on centos6
and Fedora 23 to make sure it did the right thing in both cases.

Signed-off-by: Josef Bacik <jbacik@fb.com>